+++ src/collective/cover/config.py
@@ -20,7 +20,6 @@ DEFAULT_SEARCHABLE_CONTENT_TYPES = [
     'Collection',
     'Document',
     'File',
-    'Form Folder',
     'Image',
     'Link',
     'News Item'