+++ src/collective/cover/tiles/base.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
-
 # Basic implementation taken from
 # http://davisagli.com/blog/using-tiles-to-provide-more-flexible-plone-layouts
-
 from AccessControl import Unauthorized
 from Acquisition import aq_base
 from Acquisition import aq_inner
@@ -53,6 +51,7 @@ from zope.schema import getFieldNamesInOrder
 from zope.schema import getFieldsInOrder
 
 import logging
+import Missing
 
 logger = logging.getLogger(PROJECTNAME)
 
@@ -336,6 +335,28 @@ class PersistentCoverTile(tiles.PersistentTile, ESITile):
 
         return allowed
 
+    def Date(self, brain):
+        """Return the date of publication of the object referenced by
+        brain. If the object has not been published yet, return its
+        modification date. If the object is an Event, then return the
+        start date.
+
+        :param brain: [required] brain of the cataloged object
+            referenced in the tile
+        :type brain: AbstractCatalogBrain
+        :returns: the object's publication/modification date or the
+            event's start date in case of an Event-like object
+        :rtype: str or DateTime
+        """
+        calendar = api.portal.get_tool('portal_calendar')
+        # calendar_types lists all Event-like content types
+        if brain.portal_type not in calendar.calendar_types:
+            return brain.Date
+        else:
+            # an Event must have a start date
+            assert brain.start is not Missing.Value
+            return brain.start
+
     @property
     def has_image(self):
         return self.data.get('image', None) is not None
@@ -534,7 +555,7 @@ class ImageScaling(BaseImageScaling):
         mtime = ''
         for k, v in self.context.data.items():
             if INamedImage.providedBy(v):
-                mtime += self.context.data.get('{0}_mtime'.format(k), 0)
+                mtime += self.context.data.get('{0}_mtime'.format(k), '')
 
         return mtime